refactor: scope pathfinding to drivetrain subsystem
