+++ src/main/java/com/frcteam3636/frc2024/Robot.kt
@@ -1,6 +1,5 @@
 package com.frcteam3636.frc2024
 
-import com.pathplanner.lib.pathfinding.Pathfinding
 import edu.wpi.first.hal.FRCNetComm.tInstances
 import edu.wpi.first.hal.FRCNetComm.tResourceType
 import edu.wpi.first.hal.HAL
@@ -68,7 +67,6 @@ object Robot : LoggedRobot() {
         }
 
         //TODO add field object constraints
-        Pathfinding.setPathfinder(RemoteADStarAK())
         Logger.start() // Start logging! No more data receivers, replay sources, or metadata values
         // may be added.
