+++ src/main/java/com/frcteam3636/frc2024/subsystems/drivetrain/RemoteADStarAK.java
@@ -1,4 +1,4 @@
-package com.frcteam3636.frc2024;
+package com.frcteam3636.frc2024.subsystems.drivetrain;
 
 import com.pathplanner.lib.path.GoalEndState;
 import com.pathplanner.lib.path.PathConstraints;
